requirements.txt
fastapi
uvicorn
transformers
pydantic
requests
faiss-cpu
python-dotenv
